via/requirements.txt
Django~=2.1.0
psycopg2
django-multiselectfield
django-countries
Pillow==5.2.0
django_compressor
ua-parser
